start reverting parsing changes
