+++ package.yaml
@@ -84,7 +84,6 @@ library:
     - Formula.Types
     - Formula.Resolution
     - Formula.Util
-    - Parsing
     - Util
 
   ghc-options:
