+++ src/LogicTasks/Syntax/TreeToFormula.hs
@@ -28,7 +28,7 @@ import Control.Monad (when)
 import Trees.Print (transferToPicture)
 import Tasks.TreeToFormula.Config (TreeToFormulaInst(..))
 import Formula.Parsing.Delayed (Delayed, withDelayed, parseDelayed, parseDelayedRaw)
-import Parsing (Parse(..))
+import Formula.Parsing (Parse(..))
 import Trees.Parsing()
 import UniversalParser (tokenSequence)
 import ParsingHelpers (fully)
