+++ src/Trees/Parsing.hs
@@ -20,9 +20,10 @@ import Trees.Types as Tree
     , TreeFormulaAnswer(..)
     )
 
+import Formula.Parsing(Parse(..))
+
 import ParsingHelpers (fully)
 import UniversalParser as Parser
-import Parsing (Parse(..))
 
 instance Parse (SynTree BinOp Char)
 instance FromGrammar (SynTree BinOp Char) where
