+++ test/LegalCNFSpec.hs
@@ -10,8 +10,7 @@ import Data.List((\\))
 
 import ParsingHelpers (fully)
 import Formula.Types (Cnf)
-import Parsing (parser)
-import Formula.Parsing ()
+import Formula.Parsing (parser)
 import Text.ParserCombinators.Parsec (ParseError, parse)
 import Config (CnfConfig(..), BaseConfig(..))
 import Trees.Types (SynTree(..), BinOp(..))
