+++ test/SubTreeSpec.hs
@@ -18,7 +18,7 @@ import Trees.Print (display)
 import Trees.Parsing ()
 import Trees.Types (SynTree, BinOp, PropFormula)
 import SynTreeSpec (validBoundsSynTree)
-import Parsing (Parse(parser))
+import Formula.Parsing (Parse(parser))
 import Control.Monad.Output (LangM)
 import Data.Maybe (isJust)
 import Control.Monad.Identity (Identity(runIdentity))
